all tests pass including basic web logins and nav
