+++ server/core/src/https/middleware/mod.rs
@@ -4,10 +4,12 @@ use axum::{
     response::Response,
     Extension,
 };
+use axum_sessions::SessionHandle;
 use http::{HeaderMap, HeaderValue};
 use uuid::Uuid;
 
-pub mod compression;
+pub(crate) mod caching;
+pub(crate) mod compression;
 
 // the version middleware injects
 const KANIDM_VERSION: &str = env!("CARGO_PKG_VERSION");
@@ -42,15 +44,16 @@ impl KOpId {
 /// This runs at the start of the request, adding an extension with `KOpId` which has useful things inside it.
 pub async fn kopid_start<B>(
     // TODO: try and make this into a TypedHeader - can't make it optional until at least axum 0.7.x - <https://github.com/tokio-rs/axum/issues/1781>
+    // TODO: #1787: try this https://github.com/cuberoot74088/kanidm/commit/583d7634e8d38db3ba8e980f9d4d5b64c85df849
     headers: HeaderMap,
     mut request: Request<B>,
     next: Next<B>,
 ) -> Response {
     // generate the event ID
     let eventid = sketching::tracing_forest::id();
-    // let value = eventid.as_hyphenated().to_string();
 
-    let uat = headers
+    // get the bearer token from the headers or the session
+    let uat = match headers
         .get("Authorization")
         .and_then(|hv| {
             // Get the first header value.
@@ -60,7 +63,19 @@ pub async fn kopid_start<B>(
             // Turn it to a &str, and then check the prefix
             h.strip_prefix("Bearer ")
         })
-        .map(|s| s.to_string());
+        .map(|s| s.to_string())
+    {
+        Some(val) => Some(val),
+        None => {
+            match request.extensions().get::<SessionHandle>() {
+                Some(sess) => {
+                    // we have a session!
+                    sess.read().await.get::<String>("bearer")
+                }
+                None => None,
+            }
+        }
+    };
 
     // insert the extension so we can pull it out later
     request.extensions_mut().insert(KOpId {
