+++ server/testkit/src/lib.rs
@@ -20,6 +20,10 @@ use tokio::task;
 
 pub const ADMIN_TEST_USER: &str = "admin";
 pub const ADMIN_TEST_PASSWORD: &str = "integration test admin password";
+
+pub const NOT_ADMIN_TEST_USERNAME: &str = "krab_test_user";
+pub const NOT_ADMIN_TEST_PASSWORD: &str = "eicieY7ahchaoCh0eeTa";
+
 pub static PORT_ALLOC: AtomicU16 = AtomicU16::new(18080);
 
 pub use testkit_macros::test;
